FuzzyWeight_r: float literals in interpolation expr
Apply same idiom as FuzzyWeightUndecided_r: (1.0->1.0f, double->float
cast on int division) — keeps tail mul chain at single precision.
OUR+2 -> ICM, -2 insns.

Source-commit: 99b0e91f8bd0c77596f9e1b3ada78cff2ca5f3fc

+++ botlib/botlib.c
@@ -32520,7 +32520,7 @@ double __cdecl FuzzyWeight_r(int *facts, fuzzyseperator_t *sep)
   v8 = sep->value;
   v9 = facts[sep->index] - v8;
   v10 = sep->next->value - v8;
-  return v7 * (1.0 - (double)(v9 / v10)) + (double)(v9 / v10) * v12;
+  return v7 * (1.0f - (float)(v9 / v10)) + (float)(v9 / v10) * v12;
 }
 // 10001A73: using guessed type double __cdecl FuzzyWeight_r(_DWORD, _DWORD);
